requirements.txt
mwparserfromhell>=0.6
pypandoc>=1.11
requests>=2.28
PyYAML>=6.0
